Moved string splitting to genlib/stringutils
