+++ genlib/stringutils.h
@@ -1,0 +1,28 @@
+// Copyright (C) 2017 Christian Sailer
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+// Collection of utility functions that are required to add pstring functionality
+// to std::strings (i.e. splitting and compatible serialisation/deserialisation)
+#include <vector>
+#include <string>
+#include <istream>
+#include <ostream>
+
+#pragma once
+namespace dXstring {
+    std::vector<std::string> split(const std::string &s, char delim);
+    std::string readString(const std::istream & stream);
+    void writeString(std::ostream &stream, const std::string &s);
+}
